Started cst_utils, all tests pass.
